Move getContext method out
